Refactor TwoSlider
- Pass width of slider in through props instead of getting this width
via callback.
- Clean up boundary to make more sense.
- Since the slider circle is drawn after positionX = 0, we need to
offset left slider by -1/2 of its width for it to start in middle of
the slider circle. Same for right slider, at the end of the lineWidth,
it is displaced 1 full slider width since it is drawn after
position = lineWith. Therefore, we need to offset it be -1/2 of its
width.